Add conflict detection when issuing a new lease
The algorithm used right now is very primitive and could use lots of
improvement, but it'll do for now.

We simply ARP for the IP we think is unused on the network and if we
time out, we can assume the IP is free. We choose a timeout that is a
fraction of the current context's timeout. We iterate through the
available IPs until we find one that nobody responds to.

The current code doesn't do really *any* error handling, so a legitimate
error ARP'ing may be interpreted as a timeout. We're also doing the ARP
ourselves, so we may benefit from a cache though none is implemented at
this time.